+++ dal/scopes/scope.py
@@ -12,11 +12,13 @@ Attributes:
 
 """
 from typing import List
+from functools import cached_property
 
 from dal.validation.validator import Validator
 from movai_core_shared.exceptions import DoesNotExist, AlreadyExist
 from .structures import Struct
 from dal.movaidb import MovaiDB
+from dal.movaidb.db_schema import DBSchema
 from dal.validation import JsonValidator
 
 
@@ -39,7 +41,7 @@ class Scope(Struct):
         self.__dict__["name"] = name
         self.__dict__["scope"] = scope
 
-        template_struct = MovaiDB.API()[scope]
+        template_struct = DBSchema()[scope]
 
         # we then need to get this from database!!!!
         self.__dict__["struct"] = template_struct
@@ -58,6 +60,11 @@ class Scope(Struct):
                 raise DoesNotExist(
                     f"{name} does not exist yet. If you wish to create please use 'new=True'"
                 )
+
+    @cached_property
+    def _movai_db_global(self):
+        """Instantiates the global MovaiDB object, caching it."""
+        return MovaiDB()
 
     def transform_before_update(self, source_data: dict):
         """
@@ -84,7 +91,7 @@ class Scope(Struct):
         """
         self.validate_format(self.scope, new_dict)
         structure = self.__dict__.get("struct").get("$name")
-        return MovaiDB().calc_scope_update(old_dict, new_dict, structure)
+        return self._movai_db_global.calc_scope_update(old_dict, new_dict, structure)
 
     def remove(self, force=True):
         """Removes Scope"""
@@ -93,12 +100,12 @@ class Scope(Struct):
 
     def remove_partial(self, dict_key):
         """Remove Scope key"""
-        result = MovaiDB(self.db).unsafe_delete({self.scope: {self.name: dict_key}})
+        result = self.movaidb.unsafe_delete({self.scope: {self.name: dict_key}})
         return result
 
     def get_dict(self):
         """Returns the full dictionary of the scope from db"""
-        result = MovaiDB(self.db).get({self.scope: {self.name: "**"}})
+        result = self.movaidb.get({self.scope: {self.name: "**"}})
         attrs, lists, hashs = self.get_attributes(self.struct)
         for list_name in lists:
             if list_name not in result[self.scope][self.name]: